add alignments to utils/alignment
