+++ src/main.rs
@@ -59,3 +59,4 @@ fn main() {
 
 mod processes;
 pub use crate::processes::*;
+pub(crate) mod utils;
